Estilo de acciones en vista de evento

+++ view_entry.php
@@ -940,23 +940,9 @@ if ( ( $is_my_event || $is_nonuser_admin || $is_assistant || $can_approve ) &&
     $event_status == 'W' && $readonly == 'N' && $login != '__public__') {
   $approveStr = translate ( 'Approve/Confirm entry' );
   $rejectStr = translate ( 'Reject entry' );
-  echo '<td class="styled_action">
-                <a class="action" title="' . $approveStr
-                 . '" href="approve_entry.php?id=' . $id
-                 . $u_url . '&amp;type=E" onclick="return confirm( \''
-                 . translate ( 'Approve this entry?', true ) . '\' );">
-                     <span aria-hidden="true" class="icon_check"></span>
-		&nbsp;' . $approveStr
-                 . '</a>
-	</td><td class="styled_action">
-                <a class="action" title="' . $rejectStr
-                 . '" href="reject_entry.php?id=' . $id
-                 . $u_url . '&amp;type=E" onclick="return confirm( \''
-                 . translate ( 'Reject this entry?', true ) . '\' );">
-                     <span aria-hidden="true" class="icon_close"></span>
-		&nbsp;' . $rejectStr
-                 . '</a>
-	</td>';
+  $confirm = $areYouSureStr."\\n\\n".$deleteAllStr;
+  print_action_link($approveStr, "approve_entry.php?id=" . $id . $u_url . "&amp;type=E", "icon_check", translate ( 'Approve this entry?', true ));
+  print_action_link($rejectStr, "reject_entry.php?id=" . $id . $u_url . "&amp;type=E", "icon_close",translate ( 'Reject this entry?', true ));
 }
 // TODO add these permissions to the UAC list
 $can_add_attach = ( Doc::attachmentsEnabled () && $login != '__public__'
@@ -970,23 +956,11 @@ $can_add_comment = ( Doc::commentsEnabled () && $login != '__public__'
 if ( $can_add_attach && $event_status != 'D' ) {
   $addAttchStr = translate ( 'Add Attachment' );
   
-  echo '<td class="styled_action">
-                <a class="action" title="' . $addAttchStr
-                 . '" href="docadd.php?type=A&amp;id=' . $id
-                 . $u_url . '"><span aria-hidden="true" class="icon_paperclip"></span>
-		&nbsp;' . $addAttchStr
-                 . '</a>
-	</td>';
+  print_action_link($addAttchStr, "docadd.php?type=A&amp;id=" . $id . $u_url, "icon_paperclip");
 }
 if ( $can_add_comment && $event_status != 'D' ) {
   $addCommentStr = translate ( 'Add Comment' );
-  echo '<td class="styled_action">
-                <a class="action" title="' . $addCommentStr
-                 . '" href="docadd.php?type=C&amp;id=' . $id
-                 . $u_url . '"><span aria-hidden="true" class="icon_comment_alt"></span>
-		&nbsp;' . $addCommentStr
-                 . '</a>
-	</td>';
+  print_action_link($addCommentStr, "docadd.php?type=C&amp;id=" . $id . $u_url, "icon_comment_alt");
 }
 
 // If approved, but event category not set (and user does not have permission
@@ -996,13 +970,7 @@ if ( empty ( $user ) && $CATEGORIES_ENABLED == 'Y' && $readonly != 'Y' &&
     $is_my_event && $login != '__public__' && !
     $is_nonuser && $event_status != 'D' && ! $can_edit ) {
   $setCatStr = translate ( 'Set category' );
-  echo '<td class="styled_action">
-                <a class="action" title="' . $setCatStr
-                 . '" href="set_entry_cat.php?id='
-                 . $id . $rdate . '"><span aria-hidden="true" class="icon_tag_alt"></span>
-		&nbsp;' . $setCatStr
-                 . '</a>
-	</td>';
+  print_action_link($setCatStr, "set_entry_cat.php?id=" . $id . $rdate, "icon_tag_alt");
 }
 
 $addToMineStr = translate ( 'Add to My Calendar' );
@@ -1017,44 +985,19 @@ if ( $can_edit && $event_status != 'D' && ! $is_nonuser && $readonly != 'Y' ) {
   if ( $event_repeats ) {
     $editAllDatesStr = translate ( 'Edit repeating entry for all dates' );
     $deleteAllDatesStr = translate ( 'Delete repeating event for all dates' );
-    echo '<td class="styled_action">
-                <a class="action" title="' . $editAllDatesStr
-                 . '" href="edit_entry.php?id=' . $id . $u_url . '"><span aria-hidden="true" class="icon_pencil"></span>
-		&nbsp;' . $editAllDatesStr
-                 . '</a>
-	</td>';
+    print_action_link($editAllDatesStr, "edit_entry.php?id=" . $id . $u_url, "icon_pencil");
     // Don't allow override of first event
     if ( ! empty ( $date ) && $date != $orig_date ) {
       $editThisDateStr = translate ( 'Edit entry for this date' );
-      echo '<td class="styled_action">
-                <a class="action" title="' . $editThisDateStr
-                 . '" href="edit_entry.php?id=' . $id . $u_url . $rdate . '&amp;override=1">
-                     <span aria-hidden="true" class="icon_pencil-edit_alt"></span>
-		&nbsp;' . $editThisDateStr
-                 . '</a>
-	</td>';
+      print_action_link($editThisDateStr, "edit_entry.php?id=" . $id . $u_url . $rdate . "&amp;override=1", "icon_pencil-edit_alt");
     }
-    echo '<td class="styled_action">
-                <a class="action" title="' . $deleteAllDatesStr
-                 . '" href="del_entry.php?id=' . $id . $u_url
-                 . '&amp;override=1" onclick="return confirm( \'' . $areYouSureStr . "\\n\\n"
-                 . $deleteAllStr . '\' );">
-                     <span aria-hidden="true" class="icon_trash_alt"></span>
-		&nbsp;' . $deleteAllDatesStr
-                 . '</a>
-	</td>';
+    $confirm = $areYouSureStr."\\n\\n".$deleteAllStr;
+    print_action_link($deleteAllDatesStr, "del_entry.php?id=" . $id . $u_url . "&amp;override=1", "icon_trash_alt" ,$confirm);
     // Don't allow deletion of first event
     if ( ! empty ( $date ) && $date != $orig_date ) {
       $deleteOnlyStr = translate ( 'Delete entry only for this date' );
-      echo '<td class="styled_action">
-                <a class="action" title="' . $deleteOnlyStr
-                 . '" href="del_entry.php?id='
-                    . $id . $u_url . $rdate . '&amp;override=1" onclick="return confirm( \'' . $areYouSureStr . "\\n\\n"
-                 . $deleteAllStr . '\' );">
-                     <span aria-hidden="true" class="icon_trash"></span>
-		&nbsp;' . $deleteOnlyStr
-                 . '</a>
-	</td>';
+      $confirm = $areYouSureStr."\\n\\n".$deleteAllStr;
+      print_action_link($deleteOnlyStr, "del_entry.php?id=" . $id . $u_url . $rdate . "&amp;override=1" . $id . $rdate, "icon_trash", $confirm);
     }
   } else {
     if ( ! empty( $user ) && $user != $login && ! $is_assistant ) {
@@ -1064,88 +1007,36 @@ if ( $can_edit && $event_status != 'D' && ! $is_nonuser && $readonly != 'Y' ) {
     } else {
       $delete_str = $deleteEntryStr;
     }
-    echo '<td class="styled_action">
-                <a class="action" title="' . $editEntryStr
-                 . '" href="edit_entry.php?id=' . $id . $u_url . '">
-                     <span aria-hidden="true" class="icon_pencil"></span>
-		&nbsp;' . $editEntryStr
-                 . '</a>
-	</td>
-        <td class="styled_action">
-                <a class="action" title="' . $delete_str
-                 . '" href="del_entry.php?id='
-                . $id . $u_url . $rdate . '" onclick="return confirm( \'' . $areYouSureStr
-                . "\\n\\n"
-                . ( empty ( $user ) || $user == $login || $is_assistant
-                 ? $deleteAllStr : '' )
-                . '\' );">
-                     <span aria-hidden="true" class="icon_trash"></span>
-		&nbsp;' . $delete_str
-                 . '</a>
-	</td>';
+    
+    print_action_link($editEntryStr, "edit_entry.php?id=" . $id . $u_url, "icon_pencil");
+    
+    $confirm = $areYouSureStr ."\\n\\n" . ( empty ( $user ) || $user == $login || $is_assistant ? $deleteAllStr : '' );
+    print_action_link($delete_str, "del_entry.php?id=" . $id . $u_url . $rdate , "icon_trash", $confirm);
   }
-  echo '<td class="styled_action">
-                <a class="action" title="' . $copyStr
-                 . '" href="edit_entry.php?id=' . $id . $u_url . '&amp;copy=1">
-                     <span aria-hidden="true" class="icon_document"></span>
-		&nbsp;' . $copyStr
-                 . '</a>
-	</td>';
+  
+  print_action_link($copyStr, "edit_entry.php?id=" . $id . $u_url . "&amp;copy=1", "icon_document");
   
 } elseif ( $readonly != 'Y' &&
   ( $is_my_event || $is_nonuser_admin || $can_edit ) &&
     ( $login != '__public__' ) && ! $is_nonuser && $event_status != 'D' ) {
   $delFromCalStr =
   translate ( 'This will delete the entry from your XXX calendar.', true );
-  echo '<td class="styled_action">
-                <a class="action" title="' . $deleteEntryStr
-                 . '" href="del_entry.php?id='
-   . $id . $u_url . $rdate . '" onclick="return confirm( \'' . $areYouSureStr
-   . "\\n\\n"
-   . str_replace ( 'XXX ',
-    ( $is_assistant ? translate ( 'boss' ) . ' ' : '' ), $delFromCalStr )
-  // ( $is_assistant
-  // ? translate ( 'This will delete the entry from your boss calendar.', true )
-  // : translate ( 'This will delete the entry from your calendar.', true ) )
-  . '\' );">
-                     <span aria-hidden="true" class="icon_trash"></span>
-		&nbsp;' . $deleteEntryStr
-                 . '</a>
-	</td>
-        <td class="styled_action">
-                <a class="action" title="' . $copyStr
-                 . '" href="edit_entry.php?id=' . $id . '&amp;copy=1">
-                     <span aria-hidden="true" class="icon_pencil"></span>
-		&nbsp;' . $copyStr
-                 . '</a>
-	</td>';
+  
+  $confirm = $areYouSureStr . "\\n\\n" . str_replace ( 'XXX ', ( $is_assistant ? translate ( 'boss' ) . ' ' : '' ), $delFromCalStr );
+  print_action_link($deleteEntryStr, "del_entry.php?id=" . $id . $u_url . $rdate, "icon_trash", $confirm);
+  print_action_link($copyStr, "edit_entry.php?id=" . $id . $u_url . "&amp;copy=1", "icon_document");
 }
 
 if ( $readonly != 'Y' && ! $is_my_event && ! $is_private && !
   $is_confidential && $event_status != 'D' && $login != '__public__' && !
-  $is_nonuser )
-   echo '<td class="styled_action">
-                <a class="action" title="' . $addToMineStr
-                 . '" href="add_entry.php?id='
-                . $id . '" onclick="return confirm( \''
-                . translate ( 'Do you want to add this entry to your calendar?', true )
-                . "\\n\\n" . translate ( 'This will add the entry to your calendar.', true )
-                . '\' );">
-                     <span aria-hidden="true" class="icon_document"></span>
-		&nbsp;' . $addToMineStr
-                 . '</a>
-	</td>';
+  $is_nonuser ) {
+   $confirm = translate ( 'Do you want to add this entry to your calendar?', true ) . "\\n\\n" . translate ( 'This will add the entry to your calendar.', true );
+   print_action_link($addToMineStr, "add_entry.php?id=" . $id, "icon_document", $confirm);
+  }
 
 if ( $login != '__public__' && count ( $allmails ) > 0 ) {
   $emailAllStr = translate ( 'Email all participants' );
-  echo '<td class="styled_action">
-                <a class="action" title="' . $emailAllStr
-                 . '" href="mailto:'
-   . implode ( ',', $allmails ) . '?subject=' . rawurlencode ( $subject ) . '">
-                     <span aria-hidden="true" class="icon_mail_alt"></span>
-		&nbsp;' . $emailAllStr
-                 . '</a>
-	</td>';
+  print_action_link($emailAllStr, "mailto:" . implode ( ',', $allmails ) . "?subject=" . rawurlencode ( $subject ), "icon_mail_alt");
 }
 
 $can_show_log = $is_admin; // default if access control is not enabled
@@ -1156,22 +1047,10 @@ if ( $can_show_log ) {
   $hideActivityStr = translate ( 'Hide activity log' );
   $showActivityStr = translate ( 'Show activity log' );
   if (!$show_log) {
-   echo '<td class="styled_action">
-                <a class="action" title="' . $showActivityStr
-                 . '" href="view_entry.php?id=' . $id . '&amp;log=1">
-                     <span aria-hidden="true" class="icon_ul"></span>
-		&nbsp;' . $showActivityStr
-                 . '</a>
-	</td>';
+   print_action_link($showActivityStr, "view_entry.php?id=" . $id . "&amp;log=1", "icon_ul");
       
   } else {
-      echo '<td class="styled_action">
-                <a class="action" title="' . $hideActivityStr
-                 . '" href="view_entry.php?id=' . $id . '">
-                     <span aria-hidden="true" class="icon_ul"></span>
-		&nbsp;' . $hideActivityStr
-                 . '</a>
-	</td>';
+      print_action_link($hideActivityStr, "view_entry.php?id=" . $id, "icon_ul");
   }
 }
